bin-virt/kvm-start: let the VMs start in parallel
Instead of waiting for each VM to have finished booting before starting
the next one, let's try to start all of them in such a way that they
boot in parallel.

In order to do this:
- we start a VM
- we wait for it to become 'running'
- we start the following VM
- ...

When we have started all of them, we do another pass, checking and
waiting for each VM to finish booting up to when they have SSH
available.

Chances are, especially if we are using quite a few VMs that the first
one finished booting while we were starting the others, and hence it
will be reachable via SSH right away, and we won't have to wait (and the
same for each one of the others).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>