run-kvm: add support for host config file
When running benchmarks in VMs, it may be useful to apply some special
configuration or tuning at the host level. And that may not be the same
configuration and tuning that needs to happen inside the VMs. For
example, we may want to use a specific governor at the host level, while
that may not make sense in VMs. And much more...

Let's therefore implement the support for using a specific config file
for the host. This is read and handled by run-kvm.sh and whatever is in
there is only applied at the host level.

This also avoids having to "pollute" the VM config file with the
MMTESTS_HOST_IP variable, because we can have that just in the host
config and use standard default config files from the 'configs/'
directory as they are, for VMs. Well, the VMs' config files _still_
need to have that variable defined, but if we use an host config file,
it will be run-kvm.sh itself that will add it there.

If no such host specific config file is specified, we keep using just
one config file and reading and parsing it both at the host level and
inside the VMs.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>