kvm: support defining VMs from xml config files
As of now, all the VMs in which we want to run MMTests has to
be defined (in Libvirt's sense) already. With this commit, that
is no longer a requirement.

In fact, some (or all) VMs may be "missing". In which case, we
look for a suitable set of xml configuration files that Libvirt
itself can use for defining them.

The paths where to look for such files can be specified via
both a config variable and command line. The files need to be valid
xml config files, and the names of the VMs they're defining must
match the name of the VMs we want to run the benchmarks in.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>