run-mmtests: when in VMs, stop the monitors a little earlier
If we're running inside a VM, and synchronized with the host and with
the other VMs, we wait for the proper tokens to proceed to the next step.
E.g., when we finish a test, we wait for test_done before starting
another one.

As that may, in theory, take some time, let's stop the monitor before
that, instead than after. That is, we stop the monitors closer to when
the workload that they are meant to monitor has ended. That avoids
having entries in the monitors' logs captured while the tests weren't
running any longer, and the VMs were idle.

Note that we can't do the same when it comes to start the monitors
or we risk loosing sync on the starting of the tests that don't do
multiple iterations).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>