wait_ssh_available: deal with SSH "password" prompt with small 'p'
If, for whatever reason, key login is attempted and fails (e.g., key
login is configured for the host/VM, but there are too many keys in the
agent), we'd wait forever. This situation may be problematic later but,
as far as the script is concerned, SSH is up, so we should return
success. So let's disable key login, to be sure we'll get a password
prompt.

About that, depending on OS (version), I've seen such password prompt
being "Password:" but also something like "root on 8.7.6.4 password:".
Let's try to catch that as well, by making the expect script insensitive
to the capital 'P'.

Also, while there, let's lower the backoff timeout a little. And switch
the lines to TAB indent, like the rest of the file.

Signede-off-by: Dario Faggioli <dfaggioli@suse.com>