dbench4: fetch from web instead than from git
If the mirror is out of reach, we clone the git repo. But such repo
lacks a '4.0' tag so we use master, which fails the build, with the
patches that we carry in the shellpack.

Pick the tar.gz directly, from its web location, as it still exists and
works well.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>