docs: improve wording and fix a few typos and presentation bugs
Suggested-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Cc: Mel Gorman <mgorman@suse.com>
---
Back when I sent the patches for turning the docs into markdown,
Giovanni did provide some comments (see email with msg-id:
<573000cc4c06e805783195ff24eea843b9059cbc.camel@suse.cz>).

With this patch, I'm incorporating almost all of that feedback. There's
a couple more of suggestions that were very good, but requires a bit
more of restructuring of the file, and I'm still working on those.