kvm-define: support existing VMs being undefined and refedined
In case one wants to be sure that VMs are always redefined from scratch
(e.g., from a config file), even if they existed already in Libvirt when
run-kvm.sh is invoked, this commit implements such behavior.

Whether or not this should happen is controlled (for now) only via a
(host) config variable.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>