kvm: create standalone scripts for checkinging if VMs are defined/running/etc
The one that checks if they're defined is extracted from kvm-define,
where it was a function.

While there, refactor kvm-check-running to look similar.

This commit introduces practically no functional change (except an
additional check in kvm-start) and is just preparatory.
for further changes.

While there, simplify (and harmonize) parameter handling in all
the scripts in bin-virt.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>