kvm: support automatically deploying VMs
If a VM is not defined, and we do not have an xml config file for
doing so either, let's automatically create and install it.

This commit implements the basic support for that, including
enabling some level of customization of the VM configuration as
well as choosing which Operating System should be installed
inside of it.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>