kvm: add some AutoYaST profiles, for automatic deployment
Include in the repository some AutoYaST profile, for deploying
some common openSUSE distributions inside of VMs. They're meant
to be ready to be used, but can of course be customized and/or
used as basis for new ones.

One notable thing that they do is making sure that password-based
SSH login is enabled, for the root user, in the VMs themselves,
right after install (yes, it's insecure, but we need that; and
these VMs should really have nothing to do with security
whatsoever!).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>