kvm-deploy: support using COW snapshot when importing VMs from disk images
When importing a VM, from an already existing disk image (as a
way of deploying it) a possibility is, instead of actually using
such image it as the disk of the new VM, create a COW snapshot
of it.

This (usually) saves both space and time (during the VM creation)
phase, while introducing some resource coupling/sharing, which
may (or may not!) impact performance.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>