kvm-deploy: *Initial* support for importing from predefined images
A VM can be created also importing its disk from a predefined image,
e.g., the ones downloadable from distributions' registries and official
websites.

This commit introduces the support for downloading the file from which
we import the disk of the VMs from an URL. *NOTE*, however, that some
customization of the downloaded image is most likely needed (e.g., for
interactive wizards that would trigger during first boot). This step
is not implemented yet, which means that using such images may not work.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>